Refactor Card login into CardManager
Took all card related logic from GameManager.cs and placed it into CardManager.cs
