+++ Assets/_Scripts/GameManager.cs
@@ -4,10 +4,6 @@ using UnityEngine;
 
 public class GameManager : MonoBehaviour
 {
-    public InputHandler inputHandler;
-    public PlayerMovementController playerMovementController;
-    public HandleCardStanceArrow cardStanceArrow;
-    
     public static GameManager Instance { get; private set; }
 
     private void Awake()
@@ -23,48 +19,5 @@ public class GameManager : MonoBehaviour
             Destroy(gameObject);
         }
     }
-
-    private void OnEnable()
-    {
-        // enter card stance
-        inputHandler.OnEnterCardStance += HandleEnterCardStance;
-        inputHandler.OnExitCardStance += HandleExitCardStance;
-        
-        // enable card stance directional arrow
-        inputHandler.OnEnterCardStance += HandleShowDirectionalArrow;
-        inputHandler.OnExitCardStance += HandleHideDirectionalArrow;
-    }
-
-    private void OnDisable()
-    {
-        // exit Card Stance
-        inputHandler.OnEnterCardStance -= HandleEnterCardStance;
-        inputHandler.OnExitCardStance -= HandleExitCardStance;
-        
-        // disable card stance directional arrow
-        inputHandler.OnEnterCardStance -= HandleShowDirectionalArrow;
-        inputHandler.OnExitCardStance -= HandleHideDirectionalArrow;
-    }
-
-    private void HandleEnterCardStance()
-    {
-        playerMovementController.EnterCardStance();
-    }
-
-    private void HandleExitCardStance()
-    {
-        playerMovementController.ExitCardStance();
-    }
-
-    private void HandleShowDirectionalArrow()
-    {
-       cardStanceArrow.InstantiateDirectionalArrow(); 
-    }
-
-    private void HandleHideDirectionalArrow()
-    {
-        cardStanceArrow.DestroyDirectionalArrow();
-    }
-
 }
